+++ aa_forum/tests/test_models.py
@@ -8,12 +8,13 @@ from unittest.mock import patch
 
 # Django
 from django.contrib.auth.models import Group, User
+from django.db.utils import IntegrityError
 from django.test import TestCase
 from django.urls import reverse
 from django.utils.timezone import now
 
 # AA Forum
-from aa_forum.models import Board, get_sentinel_user
+from aa_forum.models import Board, Setting, get_sentinel_user
 from aa_forum.tests.utils import (
     create_board,
     create_category,
@@ -998,14 +999,77 @@ class TestSetting(TestCase):
     Tests for Settings model
     """
 
-    def test_str(self):
+    def test_model_string_name(self):
         """
-        Test model string names
+        Test model string name
         :return:
         """
 
         # given
-        setting = create_setting(variable="alpha", value=1)
+        setting = Setting.objects.get(pk=1)
 
         # when/then
-        self.assertEqual(str(setting), "alpha")
+        self.assertEqual(str(setting), "Forum Settings")
+
+    def test_setting_save(self):
+        """
+        Test if there can't be another setting created
+        and the existing setting is changed instead
+        :return:
+        """
+
+        # given
+        setting = Setting(pk=2, default_max_messages=25, default_max_topics=25)
+        setting.save()
+
+        # then
+        self.assertEqual(setting.pk, 1)
+        self.assertEqual(setting.default_max_messages, 25)
+        self.assertEqual(setting.default_max_topics, 25)
+
+    def test_setting_create(self):
+        """
+        Test that create method throwing the following exception
+        django.db.utils.IntegrityError: (1062, "Duplicate entry '1' for key 'PRIMARY'")
+        :return:
+        """
+
+        # No pk given
+        with self.assertRaisesMessage(
+            IntegrityError, "(1062, \"Duplicate entry '1' for key 'PRIMARY'\")"
+        ):
+            create_setting()
+
+    def test_setting_create_with_pk(self):
+        """
+        Test that create method throwing the following exception no matter the given pk
+        django.db.utils.IntegrityError: (1062, "Duplicate entry '1' for key 'PRIMARY'")
+        :return:
+        """
+
+        # Set pk=2
+        with self.assertRaisesMessage(
+            IntegrityError, "(1062, \"Duplicate entry '1' for key 'PRIMARY'\")"
+        ):
+            create_setting(pk=2)
+
+    # def test_cannot_be_deleted(self):
+    #     """
+    #     Test that the settings object cannot be deleted
+    #     :return:
+    #     """
+    #
+    #     # given
+    #     settings_old = Setting.objects.get(pk=1)
+    #
+    #     # when
+    #     Setting.objects.all().delete()
+    #
+    #     # then
+    #     settings = Setting.objects.get(pk=1)
+    #
+    #     self.assertEqual(settings_old.pk, settings.pk)
+    #     self.assertEqual(
+    #         settings_old.default_max_messages, settings.default_max_messages
+    #     )
+    #     self.assertEqual(settings_old.default_max_topics, settings.default_max_topics)